Use wolfCrypt Matyas-Meyer-Oseas implementation
The faster OpenSSL implementation allocates a lot of extra memory
(equal to the size of the key) and for this reason proves to be
more erratic in performance depending on the key size and #threads.
For this reason, we prefer the use more memory efficient solutions.
In this scenario the wolfCrypt implementation proves more
performant.